Add middleware file with session validation functions and call them in routes
