+++ app.js
@@ -3,12 +3,11 @@ const path = require("path");
 const db = require("./db/database");
 const morgan = require("morgan");
 const bodyParser = require("body-parser");
-const cookieParser = require("cookie-parser");
-const crypto = require("crypto");
 const expressLayouts = require("express-ejs-layouts");
 const session = require("express-session");
 const app = express();
 const PORT = 3000;
+// TODO environment variables
 // require('dotenv').config()
 
 const indexRouter = require("./routes/index");
@@ -26,10 +25,10 @@ app.use(morgan("dev"));
 app.use("/static", express.static(path.join(__dirname, "public")));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
-// app.use(cookieParser())
 app.use(expressLayouts);
 app.use(
   session({
+    // TODO cookie saved to and deleted from db store as well
     // store: new (require('connect-pg-simple')(session))(),
     key: "user_sid",
     // TODO store secret in .env
