+++ routes/index.js
@@ -1,14 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const crypto = require("crypto");
-
-const redirectLogin = (req, res, next) => {
-  if (!req.session.userID) {
-    res.redirect("/login");
-  } else {
-    next();
-  }
-};
+const redirectLogin = require("../middleware.js").redirectLogin;
 
 router.get("/", redirectLogin, (req, res) => {
   res.redirect("/homepage");
