+++ routes/logout.js
@@ -1,13 +1,6 @@
 const express = require("express");
 const router = express.Router();
-
-const redirectLogin = (req, res, next) => {
-  if (!req.session.userID) {
-    res.redirect("/login");
-  } else {
-    next();
-  }
-};
+const redirectLogin = require("../middleware.js").redirectLogin;
 
 router.post("/", redirectLogin, (req, res) => {
   res.clearCookie("user_sid");
